add proxy retry option
If the proxy server encounters a connection refused error while sending
a request to the upstream, it will try again a certain number of times.
This accomodates cases where the upstream restarts after a file has
changed (e.g. when using the node --watch flag).

This behavior can be configured with the following options:

- UPSTREAM_RETRIES (default: 5) Number of times to retry each request.
Set to 0 to disable retrying.

- UPSTREAM_RETRY_DELAY (default: 200) Milliseconds to wait before
retrying an upstream request.